add registration closing date validation and move all exceptions to new folder
