+++ kindergarten/src/main/java/lt/vtmc/kindergarten/service/exceptions/FamilyMemberValidationException.java
@@ -1,4 +1,4 @@
-package lt.vtmc.kindergarten.service;
+package lt.vtmc.kindergarten.service.exceptions;
 
 public class FamilyMemberValidationException  extends RuntimeException {
     public FamilyMemberValidationException(String message) {
