+++ kindergarten/src/main/java/lt/vtmc/kindergarten/service/exceptions/QueueDoesntExistException.java
@@ -1,4 +1,4 @@
-package lt.vtmc.kindergarten.service;
+package lt.vtmc.kindergarten.service.exceptions;
 
 public class QueueDoesntExistException extends RuntimeException {
     public QueueDoesntExistException(String active_queue_must_exists) {
