+++ kindergarten/src/main/java/lt/vtmc/kindergarten/service/exceptions/RegistrationClosingValidationExeption.java
@@ -1,0 +1,6 @@
+package lt.vtmc.kindergarten.service.exceptions;
+
+public class RegistrationClosingValidationExeption extends RuntimeException{
+    public RegistrationClosingValidationExeption(String message) {
+    }
+}
